safari-600.7-branch/LayoutTests/fast/canvas/canvas-lineWidth-intact-after-strokeRect-expected.txt

Test that the rendering context's lineWidth is intact after calling strokeRect()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.fillStyle is '#ff0000'
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 255
PASS imgdata[4] is 0
PASS imgdata[5] is 0
PASS imgdata[6] is 0
PASS imgdata[7] is 0
PASS ctx.lineWidth is 100
PASS imgdata[0] is 0
PASS imgdata[1] is 128
PASS imgdata[2] is 0
PASS imgdata[3] is 255
PASS successfullyParsed is true

TEST COMPLETE
